server.js: Refactor the "opening a profile file" logic in the start server code (PR #4656)
